4차 스프린트 종료 백엔드 메인 브랜치 동기화

+++ backend/emm-sale/src/main/java/com/emmsale/event/application/EventService.java
@@ -1,39 +1,37 @@
 package com.emmsale.event.application;
 
-import static com.emmsale.event.exception.EventExceptionType.INVALID_MONTH;
-import static com.emmsale.event.exception.EventExceptionType.INVALID_YEAR;
-import static com.emmsale.event.exception.EventExceptionType.INVALID_YEAR_AND_MONTH;
+import static com.emmsale.event.domain.repository.EventSpecification.filterByCategory;
+import static com.emmsale.event.domain.repository.EventSpecification.filterByTags;
 import static com.emmsale.event.exception.EventExceptionType.NOT_FOUND_EVENT;
 import static com.emmsale.tag.exception.TagExceptionType.NOT_FOUND_TAG;
 import static java.util.Comparator.comparing;
 import static java.util.stream.Collectors.groupingBy;
 import static java.util.stream.Collectors.toList;
-import static java.util.stream.Collectors.toUnmodifiableList;
 
 import com.emmsale.event.application.dto.EventDetailRequest;
 import com.emmsale.event.application.dto.EventDetailResponse;
 import com.emmsale.event.application.dto.EventResponse;
-import com.emmsale.event.application.dto.ParticipantResponse;
 import com.emmsale.event.domain.Event;
 import com.emmsale.event.domain.EventStatus;
-import com.emmsale.event.domain.EventTag;
 import com.emmsale.event.domain.EventType;
-import com.emmsale.event.domain.Participant;
 import com.emmsale.event.domain.repository.EventRepository;
+import com.emmsale.event.domain.repository.EventSpecification;
 import com.emmsale.event.domain.repository.EventTagRepository;
-import com.emmsale.event.domain.repository.ParticipantRepository;
 import com.emmsale.event.exception.EventException;
 import com.emmsale.event.exception.EventExceptionType;
-import com.emmsale.member.domain.Member;
+import com.emmsale.event_publisher.EventPublisher;
 import com.emmsale.tag.application.dto.TagRequest;
 import com.emmsale.tag.domain.Tag;
 import com.emmsale.tag.domain.TagRepository;
 import com.emmsale.tag.exception.TagException;
 import java.time.LocalDate;
+import java.time.LocalDateTime;
+import java.time.format.DateTimeParseException;
 import java.util.ArrayList;
 import java.util.EnumMap;
 import java.util.List;
 import lombok.RequiredArgsConstructor;
+import org.springframework.data.jpa.domain.Specification;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
@@ -42,20 +40,13 @@ import org.springframework.transaction.annotation.Transactional;
 @RequiredArgsConstructor
 public class EventService {
 
-  private static final int MIN_MONTH = 1;
-  private static final int MAX_MONTH = 12;
-  private static final int MIN_YEAR = 2015;
+  private static final String MIN_DATE = "2000-01-01";
+  private static final String MAX_DATE = "2999-12-31";
 
   private final EventRepository eventRepository;
-  private final ParticipantRepository participantRepository;
   private final EventTagRepository eventTagRepository;
   private final TagRepository tagRepository;
-
-  private static void validateMemberNotAllowed(final Long memberId, final Member member) {
-    if (member.isNotMe(memberId)) {
-      throw new EventException(EventExceptionType.FORBIDDEN_PARTICIPATE_EVENT);
-    }
-  }
+  private final EventPublisher eventPublisher;
 
   @Transactional(readOnly = true)
   public EventDetailResponse findEvent(final Long id, final LocalDate today) {
@@ -64,141 +55,109 @@ public class EventService {
     return EventDetailResponse.from(event, today);
   }
 
-  public Long participate(final Long eventId, final Long memberId, final Member member) {
-    validateMemberNotAllowed(memberId, member);
-    final Event event = eventRepository.findById(eventId)
-        .orElseThrow(() -> new EventException(NOT_FOUND_EVENT));
-
-    final Participant participant = event.addParticipant(member);
-    participantRepository.save(participant);
-    return participant.getId();
-  }
-
-  public void cancelParticipate(final Long eventId, final Long memberId, final Member member) {
-    validateMemberNotAllowed(memberId, member);
-    if (!eventRepository.existsById(eventId)) {
-      throw new EventException(NOT_FOUND_EVENT);
-    }
-
-    participantRepository
-        .findByMemberIdAndEventId(memberId, eventId)
-        .ifPresentOrElse(
-            participant -> participantRepository.deleteById(participant.getId()),
-            () -> {
-              throw new EventException(EventExceptionType.NOT_FOUND_PARTICIPANT);
-            });
-  }
-
   @Transactional(readOnly = true)
-  public List<EventResponse> findEvents(final EventType categoryName, final LocalDate nowDate,
-      final Integer year, final Integer month,
-      final String tagName, final String statusName) {
-    List<Event> events = eventRepository.findEventsByType(categoryName);
+  public List<EventResponse> findEvents(final EventType category,
+      final LocalDate nowDate, final String startDate, final String endDate,
+      final List<String> tagNames, final List<EventStatus> statuses) {
+    Specification<Event> spec = Specification.where(filterByCategory(category));
 
-    if (isExistTagName(tagName)) {
-      events = filterByTag(tagName);
+    if (isExistTagNames(tagNames)) {
+      validateTags(tagNames);
+      spec = spec.and(filterByTags(tagNames));
     }
-    if (isExistBothYearAndMonth(year, month)) {
-      events = filterByPeriod(events, year, month);
+
+    if (isExistFilterDate(startDate, endDate)) {
+      final LocalDateTime startDateTime = validateStartDate(startDate);
+      final LocalDateTime endDateTime = validateEndDate(endDate);
+      validateEndDateAfterDateStart(startDateTime, endDateTime);
+      spec = spec.and(EventSpecification.filterByPeriod(startDateTime, endDateTime));
     }
+    final List<Event> events = eventRepository.findAll(spec);
     final EnumMap<EventStatus, List<Event>> eventsForEventStatus
         = groupByEventStatus(nowDate, events);
 
-    return filterEventResponsesByStatus(nowDate, statusName, eventsForEventStatus);
+    return filterByStatuses(nowDate, statuses, eventsForEventStatus);
   }
 
-  @Transactional(readOnly = true)
-  public List<ParticipantResponse> findParticipants(final Long eventId) {
-    final Event event = eventRepository.findById(eventId)
-        .orElseThrow(() -> new EventException(NOT_FOUND_EVENT));
-    return event.getParticipants().stream()
-        .sorted(comparing(Participant::getId))
-        .map(ParticipantResponse::from)
-        .collect(toUnmodifiableList());
+  private boolean isExistTagNames(final List<String> tagNames) {
+    return tagNames != null;
   }
 
-  private boolean isExistTagName(final String tagName) {
-    return tagName != null;
-  }
-
-  private List<Event> filterByTag(final String tagName) {
-    final Tag tag = tagRepository.findByName(tagName)
-        .orElseThrow(() -> new TagException(NOT_FOUND_TAG));
-    return eventTagRepository.findEventTagsByTag(tag)
-        .stream()
-        .map(EventTag::getEvent)
-        .collect(toList());
-  }
-
-  private boolean isExistBothYearAndMonth(final Integer year, final Integer month) {
-    if ((year == null && month != null) || (year != null && month == null)) {
-      throw new EventException(INVALID_YEAR_AND_MONTH);
-    }
-    return (year != null && month != null);
-  }
-
-  private List<Event> filterByPeriod(final List<Event> events, final int year, final int month) {
-    validateYearAndMonth(year, month);
-    return events.stream()
-        .filter(event -> isOverlapToMonth(year, month,
-            event.getStartDate().toLocalDate(), event.getEndDate().toLocalDate())
-        )
-        .collect(toList());
-  }
-
-  private void validateYearAndMonth(final int year, final int month) {
-    if (year < MIN_YEAR) {
-      throw new EventException(INVALID_YEAR);
-    }
-    if (month < MIN_MONTH || month > MAX_MONTH) {
-      throw new EventException(INVALID_MONTH);
+  private void validateTags(final List<String> tagNames) {
+    final List<Tag> tags = tagRepository.findByNameIn(tagNames);
+    if (tags.size() != tagNames.size()) {
+      throw new TagException(NOT_FOUND_TAG);
     }
   }
 
-  private boolean isOverlapToMonth(final int year, final int month,
-      final LocalDate eventStart, final LocalDate eventEnd) {
-    final LocalDate monthStart = LocalDate.of(year, month, 1);
-    final LocalDate monthEnd = LocalDate.of(year, month, monthStart.lengthOfMonth());
-
-    return (isBeforeOrEquals(eventStart, monthEnd) && isBeforeOrEquals(monthStart, eventEnd))
-        || (isBeforeOrEquals(monthStart, eventStart) && isBeforeOrEquals(eventStart, monthEnd))
-        || (isBeforeOrEquals(monthStart, eventEnd) && isBeforeOrEquals(eventEnd, monthEnd));
+  private boolean isExistFilterDate(final String startDate, final String endDate) {
+    return startDate != null || endDate != null;
   }
 
-  private boolean isBeforeOrEquals(final LocalDate criteria, final LocalDate comparison) {
-    return criteria.isBefore(comparison) || criteria.isEqual(comparison);
+  private LocalDateTime validateStartDate(final String date) {
+    try {
+      if (date == null) {
+        return LocalDate.parse(MIN_DATE).atStartOfDay();
+      }
+      return LocalDate.parse(date).atStartOfDay();
+    } catch (DateTimeParseException exception) {
+      throw new EventException(EventExceptionType.INVALID_DATE_FORMAT);
+    }
+  }
+
+  private LocalDateTime validateEndDate(final String date) {
+    try {
+      if (date == null) {
+        return LocalDate.parse(MAX_DATE).atTime(23, 59, 59);
+      }
+      return LocalDate.parse(date).atTime(23, 59, 59);
+    } catch (DateTimeParseException exception) {
+      throw new EventException(EventExceptionType.INVALID_DATE_FORMAT);
+    }
+  }
+
+  private void validateEndDateAfterDateStart(final LocalDateTime startDate,
+      final LocalDateTime endDate) {
+    if (endDate.isBefore(startDate)) {
+      throw new EventException(EventExceptionType.START_DATE_AFTER_END_DATE);
+    }
   }
 
   private EnumMap<EventStatus, List<Event>> groupByEventStatus(final LocalDate nowDate,
       final List<Event> events) {
     return events.stream()
-        .sorted(comparing(Event::getStartDate))
+        .sorted(comparing(event -> event.getEventPeriod().getStartDate()))
         .collect(
-            groupingBy(event -> event.calculateEventStatus(nowDate),
+            groupingBy(event -> event.getEventPeriod().calculateEventStatus(nowDate),
                 () -> new EnumMap<>(EventStatus.class), toList())
         );
   }
 
-
-  private List<EventResponse> filterEventResponsesByStatus(LocalDate today, final String statusName,
+  private List<EventResponse> filterByStatuses(final LocalDate today,
+      final List<EventStatus> statuses,
       final EnumMap<EventStatus, List<Event>> eventsForEventStatus) {
-    if (isExistStatusName(statusName)) {
-      EventStatus status = EventStatus.from(statusName);
-      List<Event> filteredEvents = eventsForEventStatus.get(status);
-      if (cannotFoundKeyStatus(filteredEvents)) {
-        return List.of();
-      }
-      return EventResponse.makeEventResponsesByStatus(today, status, filteredEvents);
+    if (isExistStatusName(statuses)) {
+      return filterEventResponseByStatuses(today, statuses, eventsForEventStatus);
     }
     return EventResponse.mergeEventResponses(today, eventsForEventStatus);
   }
 
-  private boolean cannotFoundKeyStatus(final List<Event> filteredEvents) {
-    return filteredEvents == null;
+  private boolean isExistStatusName(final List<EventStatus> statuses) {
+    return statuses != null;
   }
 
-  private boolean isExistStatusName(final String statusName) {
-    return statusName != null;
+  private List<EventResponse> filterEventResponseByStatuses(final LocalDate today,
+      final List<EventStatus> statuses,
+      final EnumMap<EventStatus, List<Event>> eventsForEventStatus) {
+    return eventsForEventStatus.entrySet()
+        .stream()
+        .filter(entry -> statuses.contains(entry.getKey()))
+        .map(entry -> EventResponse.makeEventResponsesByStatus(today, entry.getKey(),
+            entry.getValue()))
+        .reduce(new ArrayList<>(), (combinedEvents, eventsToAdd) -> {
+          combinedEvents.addAll(eventsToAdd);
+          return combinedEvents;
+        });
   }
 
   public EventDetailResponse addEvent(final EventDetailRequest request, final LocalDate today) {
@@ -207,6 +166,8 @@ public class EventService {
     final List<Tag> tags = findAllPersistTagsOrElseThrow(request.getTags());
 
     event.addAllEventTags(tags);
+
+    eventPublisher.publish(event);
 
     return EventDetailResponse.from(event, today);
   }
@@ -225,6 +186,8 @@ public class EventService {
         request.getLocation(),
         request.getStartDateTime(),
         request.getEndDateTime(),
+        request.getApplyStartDateTime(),
+        request.getApplyEndDateTime(),
         request.getInformationUrl(),
         tags
     );
@@ -233,8 +196,9 @@ public class EventService {
   }
 
   public void deleteEvent(final Long eventId) {
-    final Event event = eventRepository.findById(eventId)
-        .orElseThrow(() -> new EventException(NOT_FOUND_EVENT));
+    if (!eventRepository.existsById(eventId)) {
+      throw new EventException(NOT_FOUND_EVENT);
+    }
 
     eventRepository.deleteById(eventId);
   }
@@ -248,9 +212,5 @@ public class EventService {
         .map(tag -> tagRepository.findByName(tag.getName())
             .orElseThrow(() -> new EventException(EventExceptionType.NOT_FOUND_TAG)))
         .collect(toList());
-  }
-
-  public Boolean isAlreadyParticipate(final Long eventId, final Long memberId) {
-    return participantRepository.existsByEventIdAndMemberId(eventId, memberId);
   }
 }
